+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 * `Ktor`:
     * `Client`:
         * `UnifiedRequester` now have no private fields
+        * Add preview work with multipart
     * `Server`
         * `UnifiedRouter` now have no private fields
+        * Add preview work with multipart
 
 ## 0.8.6